fix : xml 바로 json으로 변환하는 방법 예시 추가
fix xml 바로 json으로 변환하는 방법 예시 추가
