+++ src/main/resources/templates/admin/festival/festivalRegulate.html
@@ -12,8 +12,6 @@
 
     <div id="festivalRegulate">
 
-        축제 관리자 페이지 입니다.
-
         <h2>축제 DB 현황</h2>
         <table class="table table-striped" id="festivalDB">
